Add missed event 2023-10-09: Beokirby deactivated
juan purported to deactivate Beokirby (among others) at [0], but this
likely failed due to the intent being malformed (intending to deactivate
multiple people in a single intent, rather than a separate intent for
each of the players). However, Beokirby was listed as inactive in the
next Registrar's report at [1], so e was ratified into being inactive at
that point. When that happened, e ceased to have a Base Rockiness switch
because Base Rockiness is an active player switch.

[0]: https://mailman.agoranomic.org/cgi-bin/mailman/private/agora-official/2023-October/017374.html
[1]: https://mailman.agoranomic.org/cgi-bin/mailman/private/agora-official/2023-October/017384.html